ref(utils): Move git related utils to their own files
