+++ static/app/utils/git/getShortCommitHash.tsx
@@ -1,0 +1,6 @@
+export function getShortCommitHash(hash: string): string {
+  if (hash.match(/^[a-f0-9]{40}$/)) {
+    hash = hash.substring(0, 7);
+  }
+  return hash;
+}
