+++ static/app/utils/git/parseRepo.spec.tsx
@@ -1,0 +1,32 @@
+import {parseRepo} from 'sentry/utils/git/parseRepo';
+
+describe('parseRepo', function () {
+  it('should work for simple github url', function () {
+    expect(parseRepo('github.com/example/example')).toEqual('example/example');
+  });
+  it('should work for full github url', function () {
+    expect(parseRepo('https://github.com/example/example')).toEqual('example/example');
+  });
+  it('should work for trailing slash', function () {
+    expect(parseRepo('https://github.com/example/example/')).toEqual('example/example');
+  });
+  it('should work for simple BitBucket url', function () {
+    expect(parseRepo('bitbucket.org/example/example')).toEqual('example/example');
+  });
+  it('should work for full BitBucket url', function () {
+    expect(parseRepo('https://bitbucket.org/example/example')).toEqual('example/example');
+  });
+  it('should work for trailing Bitbucket slash', function () {
+    expect(parseRepo('https://bitbucket.org/example/example/')).toEqual(
+      'example/example'
+    );
+  });
+  it('should work for repo only', function () {
+    expect(parseRepo('example/example')).toEqual('example/example');
+  });
+  it('should parse repo from url with extra info', function () {
+    expect(parseRepo('github.com/example/example/commits/adsadsa')).toEqual(
+      'example/example'
+    );
+  });
+});
