+++ static/app/utils/utils.spec.tsx
@@ -3,7 +3,6 @@ import {
   escapeDoubleQuotes,
   explodeSlug,
   extractMultilineFields,
-  parseRepo,
 } from 'sentry/utils';
 
 describe('utils.extractMultilineFields', function () {
@@ -27,37 +26,6 @@ three
 five`
       )
     ).toEqual(['one', 'two', 'three', 'four', 'five']);
-  });
-});
-
-describe('utils.parseRepo', function () {
-  it('should work for simple github url', function () {
-    expect(parseRepo('github.com/example/example')).toEqual('example/example');
-  });
-  it('should work for full github url', function () {
-    expect(parseRepo('https://github.com/example/example')).toEqual('example/example');
-  });
-  it('should work for trailing slash', function () {
-    expect(parseRepo('https://github.com/example/example/')).toEqual('example/example');
-  });
-  it('should work for simple BitBucket url', function () {
-    expect(parseRepo('bitbucket.org/example/example')).toEqual('example/example');
-  });
-  it('should work for full BitBucket url', function () {
-    expect(parseRepo('https://bitbucket.org/example/example')).toEqual('example/example');
-  });
-  it('should work for trailing Bitbucket slash', function () {
-    expect(parseRepo('https://bitbucket.org/example/example/')).toEqual(
-      'example/example'
-    );
-  });
-  it('should work for repo only', function () {
-    expect(parseRepo('example/example')).toEqual('example/example');
-  });
-  it('should parse repo from url with extra info', function () {
-    expect(parseRepo('github.com/example/example/commits/adsadsa')).toEqual(
-      'example/example'
-    );
   });
 });
